requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version < '3.11'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
cython==0.29.35 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
dataclasses-json==0.5.9 ; python_version >= '3.6'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
frozenlist==1.3.3 ; python_version >= '3.7'
google-search-results==2.4.2
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4 ; python_version >= '3.5'
langchain==0.0.221
langchainplus-sdk==0.0.19 ; python_version < '4.0' and python_full_version >= '3.8.1'
loguru==0.7.0 ; python_version >= '3.5'
marshmallow==3.19.0 ; python_version >= '3.7'
marshmallow-enum==1.5.1
multidict==6.0.4 ; python_version >= '3.7'
mypy-extensions==1.0.0 ; python_version >= '3.5'
numexpr==2.8.4 ; python_version >= '3.7'
numpy==1.24.4 ; python_version >= '3.8'
openai==0.27.8
openapi-schema-pydantic==1.2.4 ; python_full_version >= '3.6.1'
packaging==23.1 ; python_version >= '3.7'
pinecone-client==2.2.2
pydantic==1.10.10 ; python_version >= '3.7'
pynini==2.1.5
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pyyaml==6.0 ; python_version >= '3.6'
regex==2023.6.3 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.17 ; python_version >= '3.7'
tenacity==8.2.2 ; python_version >= '3.6'
tiktoken==0.4.0
tqdm==4.65.0 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version >= '3.7'
typing-inspect==0.9.0
urllib3==2.0.3 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
